Refactor constants out of Palette
