+++ foundry/game/gfx/objects/EnemyItemFactory.py
@@ -2,7 +2,8 @@ from PySide6.QtCore import QRect
 from PySide6.QtGui import QImage
 
 from foundry import data_dir
-from foundry.core.palette.Palette import PALETTE_GROUPS_PER_OBJECT_SET, PaletteGroup
+from foundry.core.palette import PALETTE_GROUPS_PER_OBJECT_SET
+from foundry.core.palette.Palette import PaletteGroup
 from foundry.game.gfx.drawable.Block import Block
 from foundry.game.gfx.objects.EnemyItem import EnemyObject
