+++ foundry/gui/PaletteViewer.py
@@ -1,10 +1,10 @@
 from PySide6.QtWidgets import QGridLayout, QGroupBox, QVBoxLayout
 
-from foundry.core.palette.Palette import (
+from foundry.core.palette import (
     PALETTE_GROUPS_PER_OBJECT_SET,
     PALETTES_PER_PALETTES_GROUP,
-    PaletteGroup,
 )
+from foundry.core.palette.Palette import PaletteGroup
 from foundry.game.level.LevelRef import LevelRef
 from foundry.gui.CustomDialog import CustomDialog
 from foundry.gui.PaletteWidget import PaletteWidget
